requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
httpx==0.27.0
APScheduler==3.10.4
flask==2.3.3